+++ tests/fixtures/projects/source/MultiFileMain.xs
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
 // SPDX-License-Identifier: Apache-2.0
 
-fn main() -> Long
-{
-  return add(3, 4);
+fn main() -> Long {
+    return add(3, 4);
 }