+++ xs_kts/split-example/sources/helper.xs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
 // SPDX-License-Identifier: Apache-2.0
 
-fn main() -> Long
+fn add(left: Long, right: Long) -> Long
 {
-  return add(3, 4);
+  return left + right;
 }