+++ xs_kts/split-example/xs.build.kts
@@ -4,7 +4,7 @@
  */
 
 sources {
-  include("sources/main.xs")
+  include("sources/**/*.xs")
 }
 
 compiler {